chore:done

+++ src/department-programs/department-programs.service.ts
@@ -42,12 +42,34 @@ export class DepartmentProgramsService {
   }
 
   async remove(id: number) {
-    const programToRemove =
-      await this.departmentProgramRepository.findOneById(id);
+    try {
+      // Usar findOne con opciones de consulta, incluyendo relaciones
+      const programToRemove = await this.departmentProgramRepository.findOne({
+        where: { id },
+        relations: ['Departments'], // Asegúrate de cargar la relación con Departments
+      });
 
-    if (!programToRemove) {
-      throw new NotFoundException('registro no encontrado');
+      if (!programToRemove) {
+        throw new NotFoundException('Registro no encontrado');
+      }
+
+      // Verificar si el programa tiene departamentos asociados
+      if (
+        programToRemove.Departments &&
+        programToRemove.Departments.length > 0
+      ) {
+        throw new Error(
+          'No se puede eliminar el programa porque está relacionado con departamentos',
+        );
+      }
+
+      // Eliminar el programa si no tiene departamentos asociados
+      return await this.departmentProgramRepository.remove(programToRemove);
+    } catch (error) {
+      // Controlar y lanzar la excepción apropiada
+      throw new NotFoundException(
+        error.message || 'Error interno del servidor',
+      );
     }
-    return await this.departmentProgramRepository.remove(programToRemove);
   }
 }
